backend/requirements.txt
Flask
flask-restx
flask_sqlalchemy
opencv-python
tensorflow
keras
pillow
numpy
Werkzeug